fix(355): move data to const, type file, graphql move to service
