+++ src/components/SalesDynamicsFilter/SalesDynamicsFilter.test.tsx
