+++ src/components/SalesDynamicsFilter/index.ts
@@ -1,0 +1,1 @@
+export { SalesDynamicsFilter } from './SalesDynamicsFilter';
