+++ src/constants/salesDynamics.ts
@@ -1,0 +1,5 @@
+export const GroupBy = {
+  DAY: 'DAY',
+  WEEK: 'WEEK',
+  MONTH: 'MONTH',
+} as const;
