+++ src/services/graphql/statisticsAdminService.ts
@@ -82,3 +82,23 @@ export const GET_SALES_BY_CATEGORY = gql`
     }
   }
 `;
+
+export const SALES_DYNAMICS_QUERY = gql`
+  query SalesDynamics(
+    $productIds: [ID!]!
+    $from: DateTime!
+    $to: DateTime!
+    $groupBy: GroupBy!
+  ) {
+    salesDynamics(
+      productIds: $productIds
+      from: $from
+      to: $to
+      groupBy: $groupBy
+    ) {
+      date
+      productId
+      value
+    }
+  }
+`;
